requirements.txt
-e .
aws_cdk.aws_ecs
aws_cdk.aws_ec2